maintance

+++ src/main/java/com/example/xtream/security/jwtAuthen/handler/JwtAuthenticationProvider.java
@@ -1,22 +1,17 @@
 package com.example.xtream.security.jwtAuthen.handler;
 
-import com.example.xtream.model.User;
 import com.example.xtream.service.SystemService;
 import com.example.xtream.service.TokenAuthenticationService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.Jwt;
 import io.micrometer.common.util.StringUtils;
 import lombok.RequiredArgsConstructor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 
-import java.util.Collection;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -33,7 +28,7 @@ public class JwtAuthenticationProvider extends JwtAuthenticationProviderBase {
     }
 
     @Override
-    protected List<SimpleGrantedAuthority> getAuthorities(Long userId, String userType) {
+    protected List<SimpleGrantedAuthority> getAuthorities(int userId, String userType) {
         return systemService.getAuthoritiesByUserId(userId);
     }
