+++ src/main/java/com/example/xtream/service/TokenAuthenticationService.java
@@ -4,8 +4,7 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
 
 public interface TokenAuthenticationService {
-    String createToken(Long userId, String userType);
+    String createToken(int userId, String userType);
     Boolean validateToken(String token);
     Jws<Claims> parseToken( String token);
-
 }
